Rename `testutil-time` module to `time-testlib`
