+++ buildSrc/src/main/kotlin/io/spine/dependency/local/Time.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -39,5 +39,5 @@ object Time {
     const val lib = "$group:$artifact:$version"
 
     //TODO:2024-11-29:alexander.yevsyukov: Change the artifact name to `spine-time-testlib`.
-    const val testLib = "${Spine.toolsGroup}:spine-testutil-time:$version"
+    const val testLib = "${Spine.toolsGroup}:spine-time-testlib:$version"
 }
