+++ time/build.gradle.kts
@@ -1,11 +1,11 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -26,6 +26,7 @@
 
 import io.spine.dependency.lib.AutoService
 import io.spine.dependency.local.Spine
+import io.spine.dependency.local.TestLib
 import io.spine.dependency.local.Validation
 import io.spine.gradle.publish.IncrementGuard
 
@@ -44,8 +45,8 @@ dependencies {
     api(Spine.base)
     implementation(Validation.runtime)
 
-    testImplementation(Spine.testlib)
-    testImplementation(project(":testutil-time"))
+    testImplementation(TestLib.lib)
+    testImplementation(project(":time-testlib"))
 }
 
 configurations {
